cmdlib: don't use cache qcow for composes; use virtiofs
Now that our OSBuild workflow is using the cache we saw at least one
case where the pipeline was running out of space when composing the
extensions. Since we had a previous proposal [1] to just drop the
cache altogether anyway let's try to at least remove it from the
runcompose functions to eliminate the use of it there anyway.

[1] https://github.com/coreos/coreos-assembler/pull/3615