Dockerfile: drop patching of OSBuild
The version of OSBuild in the F39 repos is newer now and the patches
don't apply any more. In the main branch we've solved this by pinning
OSBuild on specific versions. For 4.15 we don't even need to do
the patching because we aren't building using OSBuild in the pipeline
for 4.15.