cmdlib: run runcompose_extensions with a snapshot of the cache qcow
In the OSBuild workflow we'll now have a few different things using
the cache qcow at the same time. Let's run the buildextend-extensions
with a snapshot of the cache too. IIUC the artifact is copied into the
builds dir and the intermediate work done by this command isn't used
by later stages so this should be harmless.

For previous context see 6a94fe5.