osbuild: set sysroot.bootprefix true in OSTree config
This is a second round of f5677a3 (reverted in db7bf08).

Now that coreos-installer in rawhide has
https://github.com/coreos/coreos-installer/pull/1422
we can set sysroot.bootprefix to true again.

We can't do it in the non-OSBuild configs because not all streams
have the new coreos-installer yet.

Here is the comment from the original commit:

This setting will make it so that BLS config entries get prepended
with /boot. OSTree already places a boot -> . symlink in the root
of the boot filesystem prepending with /boot will always just work.

For context see https://github.com/osbuild/osbuild/issues/1566#issuecomment-1921908500

This also allows for dropping one of the upstream OSBuild zipl stage
patches.