Add `cosa supermin-run`
This is very close to `cosa supermin-shell` but instead expects a
command to run inside the supermin VM. Essentially, it's a very thin
wrapper around the `runvm` family of functions in `cmdlib.sh`.

By having it be a dedicated function, it becomes like a utility command
to have easy access to a privileged environment, with the working dir
automatically mounted in, podman remote proxying, etc... Note also this
*does not* require a cosa workdir.

The podman machine OS pipeline will be using this temporarily.