Add `cosa buildextend-layered`
As part of https://github.com/openshift/os/issues/799, we'll want to
build the "OCP node" image as a layered image on top of the RHCOS base
image.

Eventually, this image should be built outside our pipelines and more
like the rest of OpenShift container images. But for now, let's build
it ourselves. This allows us to prove out the idea without yet requiring
changes in the rest of OpenShift.

The script added here looks wordy, but it's really trivial. It's
basically a glorified wrapper around `podman build` and `skopeo copy` so
that the built OCI image ends up in our `meta.json`.