kola: handle `--oscontainer` imperatively
Right now, `--oscontainer` is handled by a systemd unit which runs
`rpm-ostree rebase` and then reboots the system. The main disadvantage
with this is that it means we don't actually get any logging of the
pivot step since it happens before SSH comes up.

But also, I want to make that switch support local OCI archives, but the
way it's currently implemented won't work with that since there's no
way to `DropFile()` it before this unit kicks in.

Move handling that arg to `runTest()` time.