kola: support passing local OCI archives to `--oscontainer`
As part of https://github.com/openshift/os/issues/799, we'll want to
be able to run tests against the layered image. We want to be able to
do that by pointing at the local file instead of having to push it to a
registry first (which in the pipeline usually happens at the end).