create_disk: run `bootupctl backend install` using bootupd from target system
Just like most tools run in this script, we want to use the version of
bootupd in the target system rather than whatever happens to be in cosa
to ensure maximum compatibility.

This is already fixed in the osbuild flow. So this is just fixing it in
the legacy flow until we've fully moved over to osbuild since it's easy
to do. While we're here, drop the `unshare` workaround since it's no
longer needed.

This bit us recently:

https://github.com/coreos/fedora-coreos-streams/issues/862#issuecomment-1992283167