osbuild: handle extra-kargs from image.yaml
I noticed that we were missing the `mitigations=auto,nosmt` from
the kernel commandline on our FCOS built images that is defined in
image-base.yaml in the FCOS configs. With OSBuild we regressed and
weren't picking up these values.

Adjust how extra kargs are passed from cmd-buildextend-metal into
create_disk.sh and runvm-osbuild so that we can support this use case.