cmd-prune: don't prune the ostree repo if non-existant
In cases where we buildfetch a bunch of different builds if we then
want to clean some of them up easily with `cosa prune` it will fail
if there is no ostree repo at `tmp/repo`. Let's just check and if
nothing is there then let's not try to clean it up.