osbuild: set sysroot.bootprefix=true for aarch64 too
We were blocked by https://github.com/coreos/fedora-coreos-tracker/issues/1647
but the 6.9 kernel series doesn't appear to have the same problem. The 6.8
kernel does have the problem but our kernel filenames in stable releases
of Fedora (i.e. not `rawhide`) won't be long enough to trigger the bug
so we should be able to safely remove this since `rawhide` has moved on
to 6.9 rc kernels.

This should be the final piece to close out
https://github.com/coreos/fedora-coreos-tracker/issues/1667