tests/basic: only run basic.nvme on x86_64 and aarch64
The original `--basic-qemu-scenarios` only ran it on x86_64. In testing,
it seems to work at least in aarch64 so we might as well test it. On
s390x, it seems like QEMU doesn't support it. On ppc64le, the bootloader
complains it's "Not a bootable device!".

Follow-up to 6775ab126 ("mantle/kola: copy basic tests into formal kola
workflow") and 2f8a427b5 ("tests/basic: fix basic tests on multi-arch").