tests/crio: pull in `crio.service` via `kubelet.service`
We default to keeping `crio.service` disabled in RHCOS. And in fact,
even in OCP `crio.service` isn't directly enabled. Instead, the
`kubelet.service` unit pulls it into the transaction as a `Requires`.

Let's mimic this setup here too by having a dummy `kubelet.service` that
pulls in `crio.service`.

The motivation for this is that this also works with the `--oscontainer`
flow, where we want to be able to run this test against a node that
booted and then pivoted to the layered ocp-node image.

And this actually highlights one of the gotchas of the layered OCP work
we'll have to keep an eye out for: it will not work to enable services
that only exist in the layered image via Ignition, because Ignition uses
presets for enablement and presets are only applied on first boot.

I think in general the way around this is to statically enable the
systemd unit instead as part of the layered image. But it just so
happens that in this particular case, we don't really need to do this
for `crio.service` anyway because even in OCP, it's not enabled via
Ignition.