tests/network: bring up machines using harness
At the time this test was written, I think `AdditionalNics` was not a
test option. So it went down the path of manually bringing up VMs with
custom QEMU options.

Nowadays, that is an option. So we can improve the logic here
by just having the harness bring up the machines for us with the
`AdditionalNics` option as required by the test.

This also makes the test compatible with `--oscontainer`, which the
harness already knows how to apply before running the test.