cmdlib.sh: go back to using `tail -F` for command output
This is more or less a revert of 4eb19f46f.

It seems like QEMU v8.2.2 (in Fedora 40) is hitting issues when
combining virtio-serial ports and the stdio character device. When the
guest writes to the virtio-serial port, it sometimes hangs.

We can look at reverting this patch if it works again in a future
version.