mantle/platform/qemu: Adjust dependencies on journal-stream unit
The unit kept getting killed when entering emergency.target, which
isn't very useful since some of the messages that hadn't been flushed
yet were the ones that revealed the root cause of the failure. Let's
set up the ordering/deps such that it doesn't get taken down when
entering emergency.target.

Also add a sleep 1 in the Destroy to allow for some remaining messages
to get flushed.