mantle/kola/testiso: add iscsi test for multipath on top of iscsi
Here we modify the iscsi setup to define two ports `:0` and `:1`
that both expose the same underlying block device, which means that
multipath can be set up on top of those two paths. We then add a
new test that provides `root=/dev/disk/by-label/dm-mpath-root` and
`rd.multipath=default` to the kernel command line to perform the test.

Note that this modifies the targetd setup and the iPXE config for all
of the tests, but they still work fine, just accessing `:0` and leaving
`:1` alone.