devshell: fix journal monitoring
Commit 7b07ffe8a dropped the `-o json` argument to the journalctl query
that the devshell code uses to monitor system boot and know when to SSH.
Fix this by re-adding the argument but only for devshell so that the
default is still to output traditional logs (as in the testiso case).

Fixes 7b07ffe8a ("mantle/platform/qemu: drop json formatting from
journal output").