osbuild: make building with OSBuild the default
Now all of our streams are ready to be switched to OSBuild and RHCOS
latest streams are also switched over. Let's make it the default so
we don't have to set COSA_USE_OSBUILD=1 any longer. This also has
the side effect of making all CI run with OSBUild now too.