osbuild: pass in ociarchive unconditionally
We'll now make deploying the OSTree directly conditional on if
ostree_repo is passed in or not. The only use case today where
we are deploying via OSTree and not via container is our FCOS
non-rawhide streams, so this shouldn't affect any non FCOS users
out there today.

This is prep for some re-work to create a buildroot from the
container. We always have the container image in our pipelines
so we want to unconditionally pass it in to OSBuild now (i.e.
so we can create the buildroot from it).