osbuild: conditionally use the built payload as the buildroot
With this we now use a buildroot that is derived from the OCI
container that was built by the pipeline. This allows us to use
the exact same versions of software from the payload we built
when we construct the images that we will ship, which will be
better for us over time.

The benefits of this are immediately apparent in this commit
as we are able to drop configuration that tries to set feature
flags for our ext4 filesystems based on what we think are the
current defaults in RHEL.

For now we aren't able to do this with FCOS because FCOS doesn't
have python in it. This should be OK for now because COSA is
almost always based on the latest version of Fedora. Though one
benefit we would have if we did switch to doing this for FCOS is
that we would test newer versions of "build tools" from `rawhide`
alongside the `rawhide` pipeline builds that we do.