docs: add section for package maintainers and contributors
We really need to make it easier for people unfamiliar with our build
processes but asked to investigate and fix some bugs to get up and
running as fast as possible. They don't need to try to learn all the
cosa intricacies and concepts.

Let's have an all-in-one section that, yes, overlaps with other
sections, but presents in a condensed way the minimum someone needs to
know to fix bugs.

Obviously, this should be shared more with fedora-bootc documentation
in the future. The ergonomics are not quite there yet however (related
discussions in https://github.com/containers/podman-bootc/pull/28 and
https://gitlab.com/fedora/bootc/tracker/-/issues/2).