kola: add support for `primaryDisk` key
Just like we have `additionalDisks`, add a new `primaryDisk` key which
takes the same "diskspec" format. This immediately unlocks all the same
knobs available to additional disks to the primary disk itself from
external tests.

E.g. with this, we can now migrate multipath tests to external tests
that use `primaryDisk: 20G:mpath`.

My immediate motivation for this however is to be able to use the
new `512e` disk option from an external test as part of fixing
OCPBUGS-35410.