cmd-sign: add `--verify-only`
This is useful to try to recover a build which was interrupted due to
signing issues. E.g. the signing request went out, but Robosignatory
never answered and the build failed.

If Robosignatory is eventually fixed and drains its queue, it eventually
does sign the build. So we then can just pick up where we left off and
verify the signatures and flip their ACLs.