kola: add IBM CEX device test for the s390x build
This kola test is crucial for verifying the security of CEX
hardware-based LUKS encryption on root volume. It guarantees that the
encrypted device employs protected keys to encrypt and decrypt the
volume.

This is essentially testing the enablement done in
https://github.com/coreos/ignition/pull/1820.

To run this, it needs to be on a system with a CEX device with
passthrough enabled and the device's UUID exposed via KOLA_CEX_UUID. See
also https://github.com/coreos/fedora-coreos-pipeline/pull/1010.

Co-authored-by: Jonathan Lebon <jonathan@jlebon.com>