cmd-buildfetch: populate `tmp/builds-source.txt` when not passing `--url`
We would write `args.url` to that file, but the user may not have passed
an explicit `--url` and instead let the default logic kick in. Instead,
just consistently use `args.url` in the fallback code so that the later
logic that writes it out will pick up on it.