kola: support relative indices for `--build` arg
I often want to target the previous build when e.g. doing `cosa run` or
`kola testiso` to compare between an older and newer build.

Instead of having to copy/paste the previous build ID, let's support
passing negative indices like `-1` for the previous build, `-2` for the
one before that, etc...

This is similar to the `journalctl --boot N` syntax.