src/cloud-prune: rework duration conversion, move to cosalib
Rework the duration parsing code to support days, weeks,
months and years. Also use a regexp so we don't need the
space.

In preparation for the container garbage collection code,
which will consume the same policy files, move the
code to the shared cosalib.

Also updates the parse_fcos_version to returns the stream id in
a tuple along the build timestamp.