cmdlib: bake src config's `live/` directory in compose
As part of the work to move the live ISO generation to osbuild and
allowing custom live ISOs, we can't have the process rely on the src
config repo. *Everything* needs to be part of the oscontainer.

The only thing `cmd-buildextend-live` needs from there is the `live/`
directory. Do like the platform stuff and shove that under
`/usr/share/coreos-assembler`.

Note we don't actually change `cmd-buildextend-live` here to copy what
it needs from there; that whole command will be gutted and rewritten to
just wrap the new osbuild code anyway.