mantle/platform/aliyun: modify the API to use HTTPS by default
Attempting to override the http scheme for each request is not working
for every request type. Let's set the client config scheme to HTTPS by default.
Move away from `NewClientWithAccessKey` and instead create an access key
separately and then use `NewClientWithOptions` to create the client with a
custom config using "HTTPS" for the scheme.
Also remove the scheme override for each request.