platform/qemu: detect if QEMU process exits unexpectedly
Currently, we only try to detect if the QEMU process exited by actually
`wait()`ing for it in the `kola qemuexec` path. We should do it in
the kola testing path as well so that it's easy to tell if e.g. it was
killed while the test was running.

Note this doesn't actually stop the test early if QEMU exited. That
would require some tricky wiring into the harness. But at least what it
prints helps diagnose the issue when we see the test time out on SSH.
And the QEMU process won't just hang there as defunct.