aliyun: make image replication idempotent
There is a history of failure copying images to other regions on
aliyun. Upstream code that calls the CopyImage code more than once
will get a DuplicateImage error when an image with the same name
already exists. Let's check to see if the image name exists in the
region before attempting to copy the image to the region and return
early if it already exists.