mantle/kola: remove rand.Seed()
This is to appease golangci-lint:

```
Error: SA1019: rand.Seed has been deprecated since Go 1.20 and an alternative
has been available since Go 1.0: As of Go 1.20 there is no reason to call Seed
with a random value. Programs that call Seed with a known value to get a
specific sequence of results should use New(NewSource(seed)) to obtain a local
random generator.  (staticcheck)
```

According to https://pkg.go.dev/math/rand#Seed "If Seed is not called, the
generator is seeded randomly at program startup." so I think it's safe to
just drop it.