Track progress - modify go routine
Since I had a problem with localizing the consolefile I changed the
code slightly to present contents of a /tmp/kola directory.
I then wrote a simple logic to eliminate errors and
output the content of a file for testing.

This will certainly be squashed, and cleaned up - for now uploading
just so I can somewhat track my progress here.