ore aws: add ability to tag images
We need the ability to tag images to facilitate garbage collection
of RHCOS AMIs that are not used in boot images. The pre-existing
tagging functionality on the initial upload cannot be leveraged because
we don't know whether a RHCOS build will be used as a boot image
at build time.