mantle/kola: simplify updates logic in kola upgrade test
Since we are dropping a config fragment down that has higher
priority than all the others that try to disable updates from
various sources (i.e. dev build, non-production build, etc) then
we don't need to worry about deleting them. Let's just let the
higher priority config win.

This also means we don't have to start and stop and start and stop
zincati so many times. Let's just start the machine with it disabled
and then start it when we've got the infra in place for it to
recognize and apply an update.

This may help avoid some race conditions now or in the future.