cmd-cloud-prune: remove the unneeded exception
This exception was not supposed to be added there since that
check was just added to make sure there isn't any build
artifacts in meta.json top-level keys that is in unsupported list.
So far, we only prune aws and gcp. There are many top-level keys
that are not artifacts-related so this check would throw the exception
when not needed.