tests/containers: add a NFS server containers
This is prep work for enabling kdump over NFS testing. The previous
attempt in https://github.com/coreos/coreos-assembler/pull/3911 used an
image from openshift E2E tests, but I didn't pay attention to the image
and the latest tag is not a multiarch manifest, so the pipeline tripped
on that.

Building the image ourselves will fix that.