simplify building the qemu-secex artifacts
We shouldn't need to pass extra parameters around for qemu-secex
and should be able to just ask for that "platform" to be produced
from OSBuild without specifying more.

This commit removes a bunch of extra variables that get set and
passed around and also refactors the osbuild manifests to have
qemu-secex treated more like the other platforms that we have in
its own qemu-secex.ipp.yaml file.

This commit also introduces a symlink cmd-buildextend-qemu-secex
that points to cmd-buildextend-metal too (in addition to the existing
cmd-buildextend-secex symlink). Since `qemu-secex` is the ID that
is used in the meta.json I think we should try to stick with it
more.