mantle/kola: add max_level_console=debug for COSA_TESTISO_DEBUG
Otherwise only `info` level messages will go to the console
of the machine.

We were digging into a new instance of [1] and found that we
weren't really getting debug messages on the console of the
machine (which means no new clues as to why the tests are
timing out). Turns out this is why we weren't getting those
new debug messages we thought we would get.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1796