plume/release: support running `update-release-index` locally
Currently, that function operates directly on an S3 bucket. This makes
it harder to test locally. Add a new `--local-mode` option in which
we expect to be in a cosa workdir with release metadata present and a
release index.

This is mostly peppering conditionals in the code. Patch best viewed
with whitespace ignored.