tests/ingition/kdump: add a remote NFS kdump test
This way we have good coverage of most-used kdump features.
Some context on the NFS kdump configuration:
coreos/fedora-coreos-tracker#1729

This was previously merged in [1] then reverted in [2] because the nfs
server container was not multi-arch, causing the pipeline to trip on it.

It's also not functionning on systemd256 (so anything f41 and above),
see [3]

This requires https://github.com/coreos/coreos-assembler/pull/3917 for
the multi-arch container, and
https://github.com/coreos/coreos-assembler/issues/3921

[1] https://github.com/coreos/coreos-assembler/commit/b10d8dcfe730d2e7c8a5366bfcddc7f3081be203
[2] https://github.com/coreos/coreos-assembler/commit/af1468c421fdedfe6c62e731eaac7fd4df720fc7
[3] https://github.com/rhkdump/kdump-utils/issues/52