osbuild: allow for multiple files to be output from a pipeline
In this case we'll just copy out the entire exported tree for
a given pipeline (which we specify as --export=platform). For
individual artifacts we'll just name the file the same name as
the platform and then copy them into the right place in the
calling script.

The benefits of this can be seen immediately because now we don't
have to know about or copy around qemu-secex bootfs_hash and
rootfs_hash inside the supermin VM.