cmdlib: only do overlayfs check in privileged path
It's only in the privileged path where we're composing directly to the
filesystem that we care about whether we're on overlayfs or not. In the
unprivileged path, the compose happens on a real filesystem on the cache
qcow2 within supermin, so there's no need to restrict what filesystem
the workdir itself on the outside is sitting on.