kola/core.go: remove fcos.internet test
This test has been denylisted in FCOS for a long time due to rate
limiting issues with the busybox container. The functionality of the
test is covered by other FCOS tests so it can be removed.
Also remove some older, obsolete functions that were probably a part
of this test in the past.

see: https://issues.redhat.com/browse/COS-2094