osbuild: rename build pipeline to deployed-tree
We're using it as more than just a buildroot now. We're more just
using it as a way to inspect and copy from or load configuration
from in various places because the files are laid out in a plain
directory structure as it would be in an OSTree deployment. Let's
rename it to more appropriately reflect what it is, which might
explain why we use it in various places.