cmdlib: check for `metadata` key before `rojig`
We've been abusing the `rojig` key for a long time now to carry artifact
name and summary information. Since then, a more appropriate `metadata
key is now supported in the treefile for our purposes.

Check the `metadata` key first, but fallback to the `rojig` key to give
time for src config repos to migrate.