Accept COSA_SUPERMIN_SIZE from environment to pass as rootfs size to supermin.
When 'supermin --build ...' generates the rootfs, kernel, and initrd, it's passed a
size to use for the rootfs.  That sets the upper bound on the ostree that can be added
to the rootfs of the images cosa generates.  It was hardcoded to 10G, which is a reasonable
starting point, but different use conditions can easily exceed that.
For now, follow the COSA_SUPERMIN_MEMORY example and expose an optional COSA_SUPERMIN_SIZE
environment variable that allows a caller to override the rootfs size in case they need the
rootfs to be bigger or smaller than the default 10G.