Stop compressing applehv and hyperv by default
Apple Hypervisor doesn't inherently require images to be compressed with
gzip. It's just that when we _do_ compress it, it's the most convenient
format to use because gzip is guaranteed to be available on macOS.

Similarly for Windows Hyper-V and ZIP.

Notably, this is different from e.g. GCP, where the platform itself
dictates a `tar.gz` file.

And so for consistency we should have the output from the build
step for `applehv` and `hyperv` just return the disk image in the
format it's intended to be used in, and then `cosa compress` just
compresses them using e.g. `gzip` or `zip`. This requires adding a new
`platform-compressor` key in the `image.yaml` file to allow overridding
the default `compressor` setting for certain platforms.

This allows folks using the same code to build the disk images for those
platforms and compress them with the compressor of their choice.