cosalib/container_manifest.py: account for images that are repo:tag
enhance the garbage collection for the repo:tag we are trying
to create a container image manifest for. In this case we were
getting an error because someone had pulled the
quay.io/coreos-assembler/coreos-assembler:main image down from
quay. It wasn't a manifest listed repo:tag so `podman manifest exists`
wasn't returning true, but the create would fail.

The updated code here should handle this case.

This enhances the previous effort in 01e4e55.