mantle/kola: use separate unit for pulling COSA in iscsi tests
We've seen errors in the RHCOS pipeline where the iSCSI tests
are hitting the 10m timeout. We could bump the timeout, but
first let's try pulling the COSA image in a separate unit so
that it can happen in parallel with the install over iSCSI
so that when we're ready to start the VM the image is already
there.