mantle/kola: make iscsi tests use virtiofs mounted COSA rootfs
In this case we'll run the podman container with --rootfs instead
of pulling a full 4+ GiB COSA image from quay for `kola qemuexec`.

This saves us quite a bit of time and bandwidth usage during a
pipeline run because we have more than 1 iscsi test.

This should also take care of https://github.com/coreos/fedora-coreos-tracker/issues/1866
because we are no longer pulling this particular container from quay.