cmd-coreos-prune: update error handling for s3.delete_object
Apparently delete object won't return a NoSuchKey error if the
object doesn't exist.

I stumbled upon this because our handling for the images action
meant we were calling prune_images() even for builds that had
already had images pruned and the logs were telling me it was
deleting the images again when it should have been telling me

```
XYZ.vmdk already pruned
```

Turns out the delete_object returns a `204 Success (No Content)`
response when the key doesn't exist.
https://docs.aws.amazon.com/AmazonS3/latest/API/API_DeleteObject.html#API_DeleteObject_RequestSyntax

Let's just delete this code in here since it's dead code and can
be confusing to future code readers.

The prior commit to this one will try to ensure we don't end up in
this code path if the images have already been pruned for a build.