image-defaults: don't set a defaults for live-rootfs*
Right now we need to ratchet in these changes using image.yaml
from the config repos rather than changing the default for
all consumers off the bat.

Note that the OSBuild stage defaults to squashfs/zstd so this is
just making it so that it what gets picked up and used for anywhere
where it's not otherwise configured.

This also has the side effect of making CI pass again because of
-zlzma,level=6 requiring a lot of memory right now:
https://github.com/erofs/erofs-utils/issues/13