cmdlib: bump cache qcow2 size
Now that we can build more and more artifacts in OSBuild it increases
the size we need for the cache qcow that gets attached to the supermin
VM. Let's bump to 40G here.

Observing a RHCOS and FCOS run here they both use right around 30G today
because FCOS builds for more platforms, but RHCOS has a bigger payload,
so they even out. In the they will be even bigger when we become able to
build the last few artifacts (vmware, virtualbox, etc) using OSBuild.