mantle/openstack: stop using boot-from-volume
Since Mantle seems to upload QCOW2 images, when using boot from
volume, it leaves a huge load on the remote control plane due
to the fact that the image needs to be converted from QCOW2 into
RAW and then booted.

The reason why this was initially changed was because boots
were taking a long time, but now OpenStack has a download-direct
from the storage backend which will do a distributed pull that
would be way faster than download-on-convert on a single system.