cmdlib.sh: limit overrides/rpm repoquery to specific arches
It's useful to be able to just `koji download-build BUILDID` to populate
`overrides/rpm` with the RPMs you want to test. Ideally you'd add e.g.
`--arch x86_64 --arch noarch` to this to limit what gets downloaded.

But it should still work fine regardless if you forgot and there's a
bunch of RPMs in there that are not for your arch. To handle this, limit
the repoquery results to the relevant arches only. This also allows us
to drop the `*.src` exclusion we had.