lib/container_manifest: check for image before manifest
There's an asymmetry between `podman image exists` and `podman manifest
exists`. The first will return 1 if the image is a manifest. The second
will return 125 if the image is _not_ a manifest.

I think this is probably a bug in podman, but for now we can work around
this by reverting the order in which we do the checking. That way, if
the image exists but isn't a manifest, we'll return `True` and not
actually call `podman manifest exists` on it.