kola/testiso: fix pxe-*.rootfs-appended test on s390x
On s390x, `mk-s390image` is used to create a bootable `pxelinux.0`
image from the live-kernel, live-initramfs, and cmdline. Because
the cmdline doesn't include `coreos.liveiso`, systemd starts the
`coreos-livepxe-rootfs.service`, which checks for the presence of
the `coreos.live.rootfs_url` karg or the `/etc/coreos-live-rootfs`
file, but neither is present. At the same time, `kola` generates
a concatenated initramfs. This PR uses that initramfs as input
for `mk-s390image`.