build.sh: add CentOS Stream keys into `/etc/pki/rpm-gpg`
This goes together with a related patch in openshift/os. Copying the
same context from that one:

A long-standing issue that rears its head in various places in our
code is the fact that the repo files for CentOS Stream reference a
`gpgkey` path that is valid only for cosa but not within a CentOS Stream
environment. See e.g. 0a7ad3b ("extensions: Workaround for CentOS GPG
key paths") in the openshift/os repo for an example issue.

We don't have this problem with RHEL because cosa, being Fedora-based,
ships the Red Hat key in its `/etc/pki/rpm-gpg`. I want to address this
for CentOS Stream the same way, i.e. by adding the CentOS Stream keys to
`/etc/pki/rpm-gpg` in cosa. This should allow us to simplify code there.