Add `cosa podman-build`
This is basically a thin wrapper around `podman build` to make it easier
to get the arguments right. The fanciest part really is the passing of
the secret repos file into the build environment.

Example usage:

```
cosa podman-build node
cosa podman-build extensions
```

Additional arguments are passed through to `podman build`.