cmdlib.sh: add support for a `fast-tracks.yaml` in src config
We often want to be able to fast-track packages from later releases
into earlier composes. But we can't just enable the whole repo because
that'll pull in *every* package.

So far, we've taken the approach of (1) asking ART to tag some
package in the plashet, and then (2) add that package under a copy
of the plashet definition in its `includepkgs=` directive so that it
effectively becomes a "RHEL-versioned" view of the plashet. The problems
with that are:
1. It's painful and slow to have to routinely tag packages via request
   tickets.
2. It's error-prone. Nothing prevents rpm-ostree from still picking an
   rhaos version of the package if it's higher NEVRA.
3. It's extremely painful to keep the production repo definitions and
   the CI definitions in sync. In practice, they just go out of skew.
4. It splits apart the definition of the packages that are being
   fast-tracked from the rest of the manifests.

Let's just teach cosa to do this for us a better way.

This patch add a new `fast-tracks.yaml` concept which can live in
the source config repo, and which can then describe which packages to
fast-tracks from which repos. cosa will then create a "filtered view" of
those repos for us that can then be used as usual in the manifests.

Example `fast-tracks.yaml` file:

```yaml
rhel-9-appstream-fast-tracks:
  from: rhel-9-appstream
  packages:
  - ignition*
```

This will define a yum repo with id `rhel-9-appstream-fast-tracks`
that's identical to `rhel-9-appstream`, but has `includepkgs=ignition*`.
(In this theoretical example, `rhel-9-appstream` is the nightly RHEL
compose still hooked to c9s. But it's equally useful to pull from e.g.
9.7 in the future for example.)

I'm not super proud of this. But I do think it's still better than what
we're currently doing.