kola/distros: introduce a `rhcos-base` variant
As we introduced pure rhel and rhcos variants in [1],
we did not change the `name` key in meta.json

Kola uses this key to determine the distribution of
the image to discriminate relevant tests.
Some tests are tied to OCP content, e.g. `crio.base`
and `crio.network` [2] and inherently won't work
with the base because they rely on content that
is added as part of the node image layer.

Currently we skip the tests tagged with `openshift`
in prow [3] and pipelie to work around this.

Going with the principle of least disruption
by adding `rhcos-base` rather than adding `rhcos-ocp`
because all of the build already out there contains
the old name and a I don't want to break compatibility
if someone use a newer cosa to test those older
builds.

[1] https://github.com/openshift/os/pull/1445
[2] https://github.com/coreos/coreos-assembler/blob/main/mantle/kola/tests/crio/crio.go
[3] https://github.com/openshift/os/blob/48a18918794f5418352c03a3415fac3fde28e1b6/ci/prow-entrypoint.sh#L306

See https://github.com/openshift/os/issues/1790